requirements.txt
# Core dependencies
groq>=0.5.0
pydantic>=2.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
langgraph>=0.0.30
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.26.0

# Type checking (optional)
mypy>=1.0.0
